+++ db/migrate/20260616150000_add_host_notified_at_to_game_days.rb
@@ -1,0 +1,5 @@
+class AddHostNotifiedAtToGameDays < ActiveRecord::Migration[7.1]
+  def change
+    add_column :game_days, :host_notified_at, :datetime
+  end
+end